Offline-Aesthetic_QR-Code_Generator/requirements.txt
qrcode[pil]
Pillow
segno
requests
numpy